+++ Countly.m
@@ -297,6 +297,7 @@ static dispatch_once_t onceToken;
     }
     CountlyContentBuilderInternal.sharedInstance.enableContentReloadOnStall = config.content.getEnableContentReloadOnStall;
     CountlyContentBuilderInternal.sharedInstance.contentReloadOnStallTimeout = config.content.getContentReloadOnStallTimeout / 1000.0;
+    CountlyContentBuilderInternal.sharedInstance.disableZoom = config.content.getDisableZoom;
 #endif
     
     [CountlyPerformanceMonitoring.sharedInstance startWithConfig:config.apm];